requirements.txt
flask
gunicorn
pymysql
torch
Pillow
transformers
requests
